test(pg_worker): move all pg_worker tests to dedicated file
Refactored tests by moving them from `tests/support/pg_worker.rs` into a new file `tests/support/pg_worker_tests.rs`. This cleans up the test support module and improves organization.
