+++ README.md
@@ -30,10 +30,10 @@ from `postgresql_embedded::Settings::default()` are used.
   PostgreSQL to stop during teardown. Defaults to `15` seconds and accepts
   values between `1` and `600`.
 
-You may also provide these values through a configuration file named
-`pg.toml`, `pg.yaml`, or `pg.json5` (depending on enabled features) located in
-any path recognised by `ortho_config`, or through CLI flags if you wrap the
-binary inside your own launcher.
+These values may also be provided through a configuration file named `pg.toml`,
+`pg.yaml`, or `pg.json5` (depending on enabled features) located in any path
+recognized by `ortho_config`, or through CLI flags when wrapping the binary
+inside a custom launcher.
 
 ## Root usage and worker binary
